fix: correct verdict-routing bugs and architecture debt from review
Fixes four confirmed problems from an architecture review, each with
regression tests:

- SARIF and JUnit output ignored PolicyFile verdict overrides, falling
  back to a kind's default policy severity instead of the canonical
  per-finding effective verdict. Both now route through
  DiffResult._effective_verdict_for_change on the override path.

- snapshot_to_dict mutated the caller's AbiSnapshot object (nulling its
  lazy lookup caches) as a side effect of serialization. It now saves
  and restores them, making the call observably pure.

- The post-processing pipeline's ctx.kept ledger relied on an implicit
  list-identity aliasing contract (established by FilterRedundant) that
  a future step could silently break by rebinding `changes` instead of
  mutating it in place. PostProcessingPipeline.run now enforces the
  contract explicitly and raises immediately if violated.

- service_scan.run_scan imported its shared scan-engine core from
  cli_scan.py, a Click command module — the reverse of the intended
  frontend -> service -> engine dependency direction. The engine core
  (run_scan_core and its helpers) is extracted into a new scan_engine.py
  with no Click dependency; both cli_scan.py and service_scan.py now
  depend on it independently. cli_scan.py also drops from 1537 to 867
  lines as a result.
