+++ tests/test_cli_contract.py
@@ -735,3 +735,57 @@ def test_run_compare_request_normalizes_lang(
     service.run_compare_request(req)
 
     assert seen_langs == ["c", "c"]
+
+
+# ── D1: service_scan must not depend on the CLI frontend ────────────────────
+#
+# service_scan.run_scan historically imported its shared scan-engine core
+# (run_scan_core / _BudgetOverflow / _EvidenceContractError) from cli_scan.py —
+# a Click command module — the reverse of the intended frontend → service →
+# engine dependency direction (ADR-037 D1). That engine core now lives in
+# scan_engine.py (no @click.option decorators, not registered as a command);
+# cli_scan.py (the CLI) and service_scan.py (the typed service API) both
+# import from it instead of service_scan reaching into the CLI module.
+
+
+def _imported_modules(path: Path) -> set[str]:
+    """Return every module name imported anywhere in *path* (module-level,
+    function-local, or under ``TYPE_CHECKING`` — all are real coupling, just
+    with different init-time consequences)."""
+    import ast
+
+    tree = ast.parse(path.read_text(encoding="utf-8"), filename=str(path))
+    out: set[str] = set()
+    for node in ast.walk(tree):
+        if isinstance(node, ast.ImportFrom) and node.module:
+            out.add(node.module)
+        elif isinstance(node, ast.Import):
+            for alias in node.names:
+                out.add(alias.name)
+    return out
+
+
+def test_service_scan_does_not_import_cli_scan() -> None:
+    """service_scan.py must never import from cli_scan.py (the Click ``scan``
+    command module) — the shared engine core lives in scan_engine.py, which
+    both cli_scan.py and service_scan.py depend on independently."""
+    import abicheck.service_scan as service_scan_mod
+
+    path = Path(service_scan_mod.__file__)
+    imported = _imported_modules(path)
+    assert "cli_scan" not in imported, (
+        "service_scan.py imports from cli_scan.py — this reintroduces the "
+        "service→CLI dependency inversion ADR-037 D1 / the scan_engine split "
+        "fixed. Import the needed symbols from abicheck.scan_engine instead."
+    )
+
+
+def test_cli_scan_reexports_the_real_scan_engine_functions() -> None:
+    """cli_scan.py's re-exported run_scan_core (etc.) are the *same objects*
+    as scan_engine's, not divergent copies — the CLI and the typed service API
+    both call one engine (ADR-037 D1)."""
+    from abicheck import cli_scan, scan_engine
+
+    assert cli_scan.run_scan_core is scan_engine.run_scan_core
+    assert cli_scan._BudgetOverflow is scan_engine._BudgetOverflow
+    assert cli_scan._EvidenceContractError is scan_engine._EvidenceContractError
